pyproject.toml
[project]
name = "solardashboard"
version = "0.1.0"
description = "Dashboard for solar data from the Fronius API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.0",
]

[tool.uv]
package = false